requirements.txt
streamlit==1.32.2

# LangChain core + integrations
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.34
langchain-groq==0.0.4
langchain-huggingface==0.0.5

# Vector store
faiss-cpu==1.7.4  # Use faiss-gpu if you have GPU setup

# Text processing and HuggingFace
huggingface-hub==0.22.2
transformers==4.39.3
torch==2.2.1  # Required for HF models
sentence-transformers==2.6.1

# PDF + file loaders
pypdf==3.17.4  # Required by PyPDFLoader
python-docx==1.1.0  # In case you later use Word files too

# Environment management
python-dotenv==1.0.1
